Add NPM Auditing for Vulnerabilities
Recently we found out that the events repo had fallen out of sync with the main
fork, and had several vulerabilities due to outdated npm dependencies.  This
commit adds a Github Action to run `npm audit` on pushes to master and on a
cron schedule. If the audit fails the Action will create an issue to alert us
to the problem.