Fix npm-audit CI bug
There was a problem in which the automatic npm dependency scanner was correctly
identifying known vulnerabilities but was not successful in creating an issue
reporting the vulnerability.

The root cause was that I copied the example from the author with insufficient
scrutiny. This example included an argument to assign the issue automatically
to `oke-py`.  Such an argument is invalid in the case of the
city-scrapers-events repo, and so it was not possible for the workflow to
create an issue. As a fix, I removed the argument from our audit.yml workflow.